+++ fixieai/cli/session/console.py
